backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
web3==6.20.1
python-dotenv==1.0.1
sqlmodel==0.0.21
requests==2.32.3
psycopg[binary]>=3.1